Update: Move pattern actions to the editor package.
