+++ packages/editor/package.json
@@ -65,6 +65,8 @@
 		"@wordpress/warning": "file:../warning",
 		"@wordpress/wordcount": "file:../wordcount",
 		"clsx": "^2.1.1",
+		"change-case": "^4.1.2",
+		"client-zip": "^2.4.4",
 		"date-fns": "^3.6.0",
 		"deepmerge": "^4.3.0",
 		"is-plain-object": "^5.0.0",
